+++ spec/punchblock/connection/xmpp_spec.rb
@@ -17,22 +17,42 @@ module Punchblock
       subject { connection }
 
       describe "rayo domains" do
+        before { stub_uuids 'randomcallid' }
+
         context "with no domains specified, and a JID of 1@app.rayo.net" do
           let(:options) { { :username => '1@app.rayo.net' } }
 
           its(:root_domain)   { should be == 'app.rayo.net' }
+
+          describe '#new_call_uri' do
+            it "should return an appropriate random call URI" do
+              subject.new_call_uri.should == 'xmpp:randomcallid@app.rayo.net'
+            end
+          end
         end
 
         context "with only a rayo domain set" do
           let(:options) { { :rayo_domain => 'rayo.org' } }
 
           its(:root_domain)   { should be == 'rayo.org' }
+
+          describe '#new_call_uri' do
+            it "should return an appropriate random call URI" do
+              subject.new_call_uri.should == 'xmpp:randomcallid@rayo.org'
+            end
+          end
         end
 
         context "with only a root domain set" do
           let(:options) { { :root_domain => 'rayo.org' } }
 
           its(:root_domain)   { should be == 'rayo.org' }
+
+          describe '#new_call_uri' do
+            it "should return an appropriate random call URI" do
+              subject.new_call_uri.should == 'xmpp:randomcallid@rayo.org'
+            end
+          end
         end
       end
 
@@ -217,6 +237,11 @@ module Punchblock
               MSG
             end
 
+            before do
+              @now = DateTime.now
+              DateTime.stub now: @now
+            end
+
             let(:example_event) { import_stanza offer_xml }
 
             it { example_event.should be_a Blather::Stanza::Presence }
@@ -228,8 +253,27 @@ module Punchblock
                 event.source_uri.should be == 'xmpp:9f00061@call.rayo.net'
                 event.domain.should be == 'call.rayo.net'
                 event.transport.should be == 'xmpp'
+                event.timestamp.should be == @now
               end
               handle_presence
+            end
+
+            context "with a delayed delivery timestamp" do
+              let :offer_xml do
+                <<-MSG
+      <presence to='16577@app.rayo.net/1' from='9f00061@call.rayo.net'>
+        <offer xmlns="urn:xmpp:rayo:1" to="sip:whatever@127.0.0.1" from="sip:ylcaomxb@192.168.1.9"/>
+        <delay xmlns='urn:xmpp:delay' stamp='2002-09-10T23:08:25Z'/>
+      </presence>
+                MSG
+              end
+
+              it 'should stamp that time on the rayo event' do
+                mock_event_handler.should_receive(:call).once.with do |event|
+                  event.timestamp.should be == DateTime.new(2002, 9, 10, 23, 8, 25, 0)
+                end
+                handle_presence
+              end
             end
           end